Optionally limit the note search by degree of separation
The note graph grows unwieldy with large number of notes, many of
which are unrelated to the current selection and clutter the view.

This patch proposes a notion of context to the query, by adding an
optional limit to the degree of separation from the active note. The
previous behaviour of the plugin is preserved by defining a max degree
of 0 to mean "retrieve all notes".